requirement.txt
#
####### example-requirements.txt #######
#
###### Requirements without Version Specifiers ######
django-extensions
imagekit
django-imagekit
django-admin-hstore-widget